Updated markdox.js to not double format when custom formatter is provided
Currently when a custom formatter is provided, the docfile is first passed through the default formatter, and then through the custom formatter. 

There might be a benefit in doing this that I'm missing, but I was assuming a formatter would have direct access to the dox output, which it doesn't right now. I also assumed that I could make a direct copy of `formatter.js` to create my own formatter, but this also doesn't work.

PS. Great project. Dox -> Markdown is exactly what I was looking for, and the templates/formatters makes this really flexible.